Create mockPanoptesResource helper Modify the dropdown editor tests to work with the generalised helper, which doesn't stub the update method.
